Minor changes to slope and predictive calculations
Added a safe division statement to the slope calculation for when no lateral movement is detected, but a new height value is received. Identical statement already existed for `timespanSeconds` and just echoed the code here as well.

Second change was to remove the additional multiplier against the `predictiveTimespan`. Incoming values were already in milliseconds and by including it here again, it resulted in reported predictive time and geometry values being off by a factor of 1000 (i.e. a user requesting a 10 seconds prediction was receiving statistics calculated against 10000 seconds).